reducing ember version support to match ember-animated
