+++ app/initializers/liquid-fire.js
@@ -1,7 +1,4 @@
-import { initialize } from "liquid-fire/ember-internals";
 import "liquid-fire/velocity-ext";
-
-initialize();
 
 export default {
   name: 'liquid-fire',
