+++ config/ember-try.js
@@ -11,6 +11,27 @@ module.exports = function() {
     return {
       scenarios: [
         {
+          name: 'ember-lts-2.12',
+          npm: {
+            devDependencies: {
+              'ember-source': '~2.12.0',
+              'ember-native-dom-event-dispatcher': '~0.6.4',
+            }
+          }
+        },
+        {
+          name: 'ember-lts-2.16',
+          env: {
+            EMBER_OPTIONAL_FEATURES: JSON.stringify({ 'jquery-integration': true })
+          },
+          npm: {
+            devDependencies: {
+              'ember-source': '~2.16.0',
+              'ember-native-dom-event-dispatcher': '~0.6.4',
+            }
+          }
+        },
+        {
           name: 'ember-lts-2.18',
           env: {
             EMBER_OPTIONAL_FEATURES: JSON.stringify({ 'jquery-integration': true })
@@ -62,19 +83,6 @@ module.exports = function() {
           name: 'ember-default',
           npm: {
             devDependencies: {}
-          }
-        },
-        {
-          name: 'ember-default-with-jquery',
-          env: {
-            EMBER_OPTIONAL_FEATURES: JSON.stringify({
-              'jquery-integration': true
-            })
-          },
-          npm: {
-            devDependencies: {
-              '@ember/jquery': '^0.5.1'
-            }
           }
         }
       ]
