+++ package.json
@@ -22,8 +22,6 @@
     "ember-cli-htmlbars": "^3.0.1",
     "ember-cli-version-checker": "^3.1.3",
     "ember-copy": "^1.0.0",
-    "ember-getowner-polyfill": "^2.2.0",
-    "ember-hash-helper-polyfill": "^0.2.1",
     "match-media": "^0.2.0",
     "velocity-animate": "^1.5.2"
   },
