--- version-specific-1.13/get-outlet-state.hbs
@@ -1,3 +1,0 @@
-{{#-get-outlet-state as |o|~}}
-  {{yield o}}
-{{~/-get-outlet-state}}
