--- version-specific-2.9/index.js
@@ -1,11 +1,0 @@
-import Ember from 'ember';
-
-const { getViewBounds } = Ember.ViewUtils;
-
-export function initialize() { }
-
-export function containingElement(view) {
-  return getViewBounds(view).parentElement;
-}
-
-export { default as getOutletStateTemplate } from 'liquid-fire/templates/version-specific/get-outlet-state';
